Pipelines/mek_photos_to_gs/requirements.txt
pymongo>=3.12.0,<=3.20.0
google-cloud-storage>=1.42.0,<=1.50.0
python-decouple==3.4
